Conflicts:
	deps/boost/boost/config/compiler/clang.hpp
	deps/boost/boost/uuid/uuid_io.hpp

+++ deps/boost/libs/asio/example/cpp03/chat/chat_message.hpp
@@ -2,7 +2,7 @@
 // chat_message.hpp
 // ~~~~~~~~~~~~~~~~
 //
-// Copyright (c) 2003-2013 Christopher M. Kohlhoff (chris at kohlhoff dot com)
+// Copyright (c) 2003-2014 Christopher M. Kohlhoff (chris at kohlhoff dot com)
 //
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -81,7 +81,7 @@ public:
   {
     using namespace std; // For sprintf and memcpy.
     char header[header_length + 1] = "";
-    sprintf(header, "%4d", body_length_);
+    sprintf(header, "%4d", static_cast<int>(body_length_));
     memcpy(data_, header, header_length);
   }
