+++ deps/boost/libs/range/test/iterator_pair.cpp
@@ -63,18 +63,24 @@ void check_iterator_pair()
     BOOST_CHECK( boost::begin( pair ) == pair.first );
     BOOST_CHECK( boost::end( pair )   == pair.second );
     BOOST_CHECK( boost::empty( pair ) == (pair.first == pair.second) );
-    BOOST_CHECK( boost::size( pair )  == std::distance( pair.first, pair.second ) );
+    BOOST_CHECK( boost::size( pair )  ==
+                    static_cast<std::size_t>(
+                        std::distance(pair.first, pair.second)) );
 
     BOOST_CHECK( boost::begin( const_pair ) == const_pair.first );
     BOOST_CHECK( boost::end( const_pair )   == const_pair.second );
     BOOST_CHECK( boost::empty( const_pair ) == (const_pair.first == const_pair.second) );
-    BOOST_CHECK( boost::size( const_pair )  == std::distance( const_pair.first, const_pair.second ) );
+    BOOST_CHECK( boost::size( const_pair )  ==
+                    static_cast<std::size_t>(
+                        std::distance(const_pair.first, const_pair.second)) );
 
     BOOST_CHECK( boost::begin( constness_pair ) == constness_pair.first );
     BOOST_CHECK( boost::end( constness_pair )   == constness_pair.second );
     BOOST_CHECK( boost::empty( constness_pair ) == (constness_pair.first == const_pair.second) );
-    BOOST_CHECK( boost::size( constness_pair )  == std::distance( constness_pair.first, constness_pair.second ) );
-
+    BOOST_CHECK( boost::size( constness_pair )  ==
+                    static_cast<std::size_t>(
+                        std::distance(constness_pair.first,
+                                      constness_pair.second)) );
 }
